Clean up unit tests
